Refactoring directory structute to make more meaningful sense
